.skeleton/requirements.txt
jinja2
pyyaml
gitpython
